adds configured db
