+++ backend/package.json
@@ -7,7 +7,7 @@
     "dev": "cross-env NODE_ENV=development npx webpack --mode development",
     "build": "cross-env NODE_ENV=production npx webpack --mode production",
     "watch": "cross-env NODE_ENV=development npx webpack --mode development --watch",
-    "start": "cross-env NODE_ENV=development webpack-dev-server --mode development --open"    
+    "start": "cross-env NODE_ENV=development webpack-dev-server --mode development --open"
   },
   "browsersList": ">0.25%, not dead",
   "repository": {
@@ -38,6 +38,7 @@
     "mini-css-extract-plugin": "^2.5.3",
     "node-sass": "^7.0.1",
     "optimize-css-assets-webpack-plugin": "^6.0.1",
+    "prisma": "^3.10.0",
     "sass-loader": "^12.4.0",
     "style-loader": "^3.3.1",
     "terser-webpack-plugin": "^5.3.0",
@@ -48,6 +49,7 @@
   },
   "dependencies": {
     "@babel/polyfill": "^7.12.1",
+    "@prisma/client": "^3.10.0",
     "express": "^4.17.3",
     "express-handlebars": "^6.0.2",
     "handlebars-loader": "^1.7.1",
