+++ backend/prisma/schema.prisma
@@ -1,0 +1,26 @@
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+}
+
+model Products {
+  title                 String  @db.VarChar(50)
+  price                 String  @db.VarChar(50)
+  old_price             String? @db.VarChar(50)
+  category              String  @db.VarChar(50)
+  description           String
+  image_url             String  @db.VarChar(500)
+  raiting               Int
+  quantity              Int?
+  isNew                 Boolean
+  isUltraSale           Boolean
+  isFeatureProduct      Boolean
+  id                    Int     @id(map: "products_pkey") @default(autoincrement())
+  additionalDescription String?
+  additionalInfo        String?
+  reviews               String?
+}
