--- backend/products/test.js
@@ -1,1542 +1,0 @@
-module.exports = [
-  {
-    id: 43,
-    title: 'Food Colouring - Blue',
-    price: '$26.69',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 5,
-    quantity: 20,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 12,
-    title: 'Beer - Camerons Auburn',
-    price: '$6.15',
-    old_price: null,
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 3,
-    quantity: 2,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 74,
-    title: 'Wine - Vineland Estate Semi - Dry',
-    price: '$16.28',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 5,
-    quantity: 16,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 28,
-    title: 'Rum - White, Gg White',
-    price: '$28.82',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 3,
-    quantity: 2,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 29,
-    title: 'Iced Tea Concentrate',
-    price: '$12.51',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 3,
-    quantity: 4,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 5,
-    title: 'Bread - Sour Batard',
-    price: '$27.72',
-    old_price: '$32.72',
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 4,
-    quantity: 20,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 49,
-    title: 'Table Cloth 53x53 White',
-    price: '$10.09',
-    old_price: '$15.09',
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 5,
-    quantity: 19,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 47,
-    title: 'Alize Gold Passion',
-    price: '$16.83',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 5,
-    quantity: 12,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 95,
-    title: 'Cheese - Brick With Pepper',
-    price: '$23.12',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 3,
-    quantity: 7,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 96,
-    title: 'Cod - Black Whole Fillet',
-    price: '$11.79',
-    old_price: '$17.99',
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 3,
-    quantity: 4,
-    isNew: true,
-    isUltraSale: true,
-    isFeatureProduct: false,
-  },
-  {
-    id: 85,
-    title: 'Amaretto',
-    price: '$12.49',
-    old_price: '$17.49',
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 10,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 21,
-    title: 'Bols Melon Liqueur',
-    price: '$16.13',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 5,
-    quantity: 5,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 94,
-    title: 'Pepsi, 355 Ml',
-    price: '$19.19',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 13,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 87,
-    title: 'Venison - Denver Leg Boneless',
-    price: '$29.97',
-    old_price: '$34.97',
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 13,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 63,
-    title: 'Sauce - Demi Glace',
-    price: '$29.44',
-    old_price: '$36.66',
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 5,
-    quantity: 11,
-    isNew: false,
-    isUltraSale: true,
-    isFeatureProduct: false,
-  },
-  {
-    id: 18,
-    title: 'Chocolate - Feathers',
-    price: '$14.64',
-    old_price: null,
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 4,
-    quantity: 12,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 42,
-    title: 'Bread - Raisin Walnut Pull',
-    price: '$14.25',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 3,
-    quantity: 9,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 100,
-    title: 'Puree - Mango',
-    price: '$23.72',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 5,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 34,
-    title: 'Wine - Zinfandel California 2002',
-    price: '$7.06',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 4,
-    quantity: 8,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 93,
-    title: 'Ecolab Silver Fusion',
-    price: '$13.75',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 3,
-    quantity: 20,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 26,
-    title: 'Table Cloth 90x90 Colour',
-    price: '$20.81',
-    old_price: '$25.81',
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 3,
-    quantity: 3,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 53,
-    title: 'Wasabi Powder',
-    price: '$8.98',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 4,
-    quantity: 10,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 98,
-    title: 'Veal Inside - Provimi',
-    price: '$26.22',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 3,
-    quantity: 4,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 55,
-    title: 'Bread - Pumpernickel',
-    price: '$14.15',
-    old_price: '$19.15',
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 3,
-    quantity: 2,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 23,
-    title: 'Filter - Coffee',
-    price: '$27.96',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 3,
-    quantity: 15,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 80,
-    title: 'Stock - Chicken, White',
-    price: '$14.30',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 3,
-    quantity: 5,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 69,
-    title: 'Radish - Black, Winter, Organic',
-    price: '$5.40',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 4,
-    quantity: 9,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 66,
-    title: 'Curry Paste - Madras',
-    price: '$25.85',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 3,
-    quantity: 15,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 61,
-    title: 'Bar - Sweet And Salty Chocolate',
-    price: '$11.46',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 4,
-    quantity: 3,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 32,
-    title: 'Raisin - Dark',
-    price: '$11.41',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 5,
-    quantity: 16,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 59,
-    title: 'Oil - Canola',
-    price: '$18.87',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 4,
-    quantity: 6,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 82,
-    title: 'Sauce - Caesar Dressing',
-    price: '$14.85',
-    old_price: '$19.85',
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 19,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 99,
-    title: 'Bay Leaf',
-    price: '$16.36',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 4,
-    quantity: 1,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 4,
-    title: 'Melon - Cantaloupe',
-    price: '$6.66',
-    old_price: '$13.86',
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 4,
-    quantity: 17,
-    isNew: true,
-    isUltraSale: true,
-    isFeatureProduct: false,
-  },
-  {
-    id: 7,
-    title: 'Sweet Apples',
-    price: '$20.24',
-    old_price: '$27.46',
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 4,
-    quantity: 7,
-    isNew: false,
-    isUltraSale: true,
-    isFeatureProduct: false,
-  },
-  {
-    id: 8,
-    title: 'Beer - Sleemans Honey Brown',
-    price: '$11.53',
-    old_price: null,
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 4,
-    quantity: 8,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 83,
-    title: 'Salad Dressing',
-    price: '$20.29',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 11,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 60,
-    title: 'Dried Cranberries',
-    price: '$7.76',
-    old_price: '$12.76',
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 3,
-    quantity: 9,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 52,
-    title: 'Veal - Eye Of Round',
-    price: '$7.47',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 4,
-    quantity: 2,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 78,
-    title: 'Wine - White, Riesling, Semi - Dry',
-    price: '$15.32',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 3,
-    quantity: 6,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 36,
-    title: 'Veal - Kidney',
-    price: '$14.30',
-    old_price: '$19.3',
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 4,
-    quantity: 20,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 75,
-    title: 'Sauce - Plum',
-    price: '$10.78',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 3,
-    quantity: 16,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 46,
-    title: 'Spic And Span All Purpose',
-    price: '$13.54',
-    old_price: '$18.54',
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 4,
-    quantity: 10,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 79,
-    title: 'Flavouring Vanilla Artificial',
-    price: '$24.28',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 5,
-    quantity: 20,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 25,
-    title: 'Apron',
-    price: '$10.98',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 3,
-    quantity: 16,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 15,
-    title: 'Bread - White Mini Epi',
-    price: '$17.12',
-    old_price: null,
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 4,
-    quantity: 20,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 90,
-    title: 'Aspic - Amber',
-    price: '$17.59',
-    old_price: '$22.59',
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 8,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 31,
-    title: 'Arctic Char - Fillets',
-    price: '$23.73',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 3,
-    quantity: 16,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 30,
-    title: 'Chocolate - Pistoles, White',
-    price: '$16.34',
-    old_price: '$21.34',
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 4,
-    quantity: 19,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 38,
-    title: 'Coffee - Decaffeinato Coffee',
-    price: '$5.04',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 5,
-    quantity: 20,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 14,
-    title: 'Muffins - Assorted',
-    price: '$6.83',
-    old_price: '$11.83',
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 4,
-    quantity: 2,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 48,
-    title: 'Foam Espresso Cup Plain White',
-    price: '$22.34',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 4,
-    quantity: 8,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 62,
-    title: 'Bread - White Mini Epi',
-    price: '$24.15',
-    old_price: '$29.15',
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 4,
-    quantity: 11,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 76,
-    title: 'Chocolate - Sugar Free Semi Choc',
-    price: '$18.51',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 5,
-    quantity: 18,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 10,
-    title: 'Chinese Foods - Cantonese',
-    price: '$21.59',
-    old_price: null,
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 5,
-    quantity: 19,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 45,
-    title: 'Lambcasing',
-    price: '$22.13',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 3,
-    quantity: 2,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 19,
-    title: 'Flounder - Fresh',
-    price: '$13.21',
-    old_price: '$18.21',
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 4,
-    quantity: 7,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 71,
-    title: 'Goat - Leg',
-    price: '$18.96',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 5,
-    quantity: 20,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 24,
-    title: 'Wine - Red, Concha Y Toro',
-    price: '$28.85',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 4,
-    quantity: 7,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 51,
-    title: 'Pork - Suckling Pig',
-    price: '$6.81',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 5,
-    quantity: 17,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 1,
-    title: 'Mix - Cocktail Ice Cream',
-    price: '$15.07',
-    old_price: null,
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 3,
-    quantity: 7,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 2,
-    title: 'Tea - Vanilla Chai',
-    price: '$5.69',
-    old_price: '$10.690000000000001',
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 5,
-    quantity: 7,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 44,
-    title: 'Wine - Magnotta - Bel Paese White',
-    price: '$23.69',
-    old_price: '$28.69',
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 4,
-    quantity: 5,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 3,
-    title: 'Eggplant - Regular',
-    price: '$27.64',
-    old_price: '$32.64',
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 3,
-    quantity: 17,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 92,
-    title: 'Lemon Pepper',
-    price: '$28.81',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 12,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 33,
-    title: 'Cod - Black Whole Fillet',
-    price: '$6.38',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 5,
-    quantity: 8,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 77,
-    title: 'Wine - White, Pinot Grigio',
-    price: '$19.94',
-    old_price: '$24.94',
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 3,
-    quantity: 14,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 56,
-    title: 'Apple - Granny Smith',
-    price: '$26.03',
-    old_price: '$31.03',
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 5,
-    quantity: 2,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 11,
-    title: 'Lettuce - Romaine, Heart',
-    price: '$12.39',
-    old_price: null,
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 5,
-    quantity: 15,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 73,
-    title: 'Wine - Peller Estates Late',
-    price: '$11.22',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 4,
-    quantity: 8,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 91,
-    title: 'Grouper - Fresh',
-    price: '$9.72',
-    old_price: '$14.72',
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 3,
-    quantity: 18,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 88,
-    title: 'Spic And Span All Purpose',
-    price: '$20.51',
-    old_price: '$25.51',
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 4,
-    quantity: 2,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 72,
-    title: 'Wine - Fontanafredda Barolo',
-    price: '$20.68',
-    old_price: '$25.68',
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 4,
-    quantity: 10,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 39,
-    title: 'Wine - White, Ej',
-    price: '$16.83',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 4,
-    quantity: 4,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 6,
-    title: 'Bacardi Limon',
-    price: '$22.28',
-    old_price: null,
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 4,
-    quantity: 8,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 22,
-    title: 'Lemonade - Black Cherry, 591 Ml',
-    price: '$11.13',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 5,
-    quantity: 12,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 57,
-    title: 'Salad Dressing',
-    price: '$21.60',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 4,
-    quantity: 3,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 27,
-    title: 'Club Soda - Schweppes, 355 Ml',
-    price: '$16.91',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 4,
-    quantity: 2,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 97,
-    title: 'Peas - Frozen',
-    price: '$10.41',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 3,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 9,
-    title: 'Sugar - Fine',
-    price: '$6.86',
-    old_price: '$11.86',
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 5,
-    quantity: 15,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 35,
-    title: 'Vinegar - White',
-    price: '$17.24',
-    old_price: '$22.24',
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 5,
-    quantity: 7,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 13,
-    title: 'Milk - Homo',
-    price: '$28.57',
-    old_price: null,
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 3,
-    quantity: 2,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 68,
-    title: 'Cinnamon Rolls',
-    price: '$14.87',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 5,
-    quantity: 17,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 81,
-    title: 'Basil - Seedlings Cookstown',
-    price: '$29.95',
-    old_price: '$34.95',
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 5,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 41,
-    title: 'Salt - Seasoned',
-    price: '$5.08',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 4,
-    quantity: 14,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 17,
-    title: 'Wine - Prosecco Valdobienne',
-    price: '$16.48',
-    old_price: '$21.48',
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 5,
-    quantity: 7,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 89,
-    title: 'Blueberries',
-    price: '$10.71',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 16,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 65,
-    title: 'Salt - Kosher',
-    price: '$18.72',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 4,
-    quantity: 17,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 16,
-    title: 'Fish - Soup Base, Bouillon',
-    price: '$29.12',
-    old_price: '$34.120000000000005',
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 4,
-    quantity: 14,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 40,
-    title: 'Lobak',
-    price: '$17.33',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 3,
-    quantity: 2,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 70,
-    title: 'Dill - Primerba, Paste',
-    price: '$27.56',
-    old_price: '$32.56',
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 4,
-    quantity: 15,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 37,
-    title: 'Onions - Cippolini',
-    price: '$10.26',
-    old_price: null,
-    category: 'vegetables',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://befreshcorp.net/wp-content/uploads/2017/07/product-packshot-aubergine-558x600.jpg',
-    raiting: 5,
-    quantity: 9,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: true,
-  },
-  {
-    id: 54,
-    title: 'Onions - Vidalia',
-    price: '$17.95',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 5,
-    quantity: 16,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 58,
-    title: 'Wine - Penfolds Koonuga Hill',
-    price: '$15.50',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 5,
-    quantity: 19,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 84,
-    title: 'Chicken Breast Wing On',
-    price: '$27.63',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 3,
-    quantity: 16,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 86,
-    title: 'Shrimp - 31/40',
-    price: '$8.76',
-    old_price: null,
-    category: 'smoothies',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://st3.depositphotos.com/16122460/31657/i/450/depositphotos_316574936-stock-photo-tasty-fresh-milk-shake-in.jpg',
-    raiting: 5,
-    quantity: 4,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-  {
-    id: 67,
-    title: 'Wine - Cousino Macul Antiguas',
-    price: '$5.65',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 5,
-    quantity: 8,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: true,
-  },
-  {
-    id: 64,
-    title: 'Plate - Foam, Bread And Butter',
-    price: '$28.68',
-    old_price: null,
-    category: 'grainAndNuts',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://vkustorg.com/image/cache/Sukhofrukty/44301251-1000x1000.jpg',
-    raiting: 5,
-    quantity: 20,
-    isNew: false,
-    isUltraSale: false,
-    isFeatureProduct: true,
-  },
-  {
-    id: 20,
-    title: 'Oven Mitt - 13 Inch',
-    price: '$10.45',
-    old_price: null,
-    category: 'fruits',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url: 'https://m.dom-eda.com/uploads/images/catalog/item/86df51de21/c25c94fe96_1000.jpg',
-    raiting: 5,
-    quantity: 7,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: true,
-  },
-  {
-    id: 50,
-    title: 'V8 - Berry Blend',
-    price: '$16.01',
-    old_price: null,
-    category: 'meats',
-    description:
-      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.",
-    image_url:
-      'https://thumbs.dreamstime.com/b/%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D1%8B%D0%B5-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B0-%D0%B8%D0%B7%D0%BE%D0%BB%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD%D0%BD%D0%BE%D0%B9-%D0%BD%D0%B0-%D0%B1%D0%B5%D0%BB%D0%BE%D0%BC-%D1%84%D0%BE%D0%BD%D0%B5-raw-%D1%84%D0%B8%D0%BB%D0%B5-%D0%BA%D1%83%D1%80%D0%B8%D0%BD%D0%BE%D0%B9-%D0%B3%D1%80%D1%83%D0%B4%D0%BA%D0%B8-172873957.jpg',
-    raiting: 5,
-    quantity: 16,
-    isNew: true,
-    isUltraSale: false,
-    isFeatureProduct: false,
-  },
-];
